separate mock command station
